Cleanup, rename client components `.client.tsx`
